WIP checkpoint: experiments, faithfulness writeup, ELBO reference notebook
Snapshot of all working-tree experimental work so it's accessible remotely
(mobile). Includes:
- notebooks/ELBO_for_DBN.ipynb + builder: clean from-scratch faithful reference
  implementation of the bar-pointer VAE (strict ELBO, no bandages)
- FAITHFULNESS.md writeup
- tests/: joint eval, SMC/PF eval, DBN baselines, fast-proxy harness, unit tests
- models/bar_pointer_dbn.py, distributions/loss tweaks
- run_*.sh experiment drivers
- gitignore /cache/ (~18GB activations + diag checkpoints; regenerable)

+++ tests/pf_eval_cache.py
@@ -45,7 +45,7 @@ def main() -> int:
     with torch.no_grad():
         for f in fs:
             r = torch.load(f, map_location=dev)
-            act = r["activations"][:maxT].unsqueeze(0)
+            act = r["activations"][:maxT].unsqueeze(0).float()   # fp16 rich cache -> fp32
             fps = float(r["fps"])
             ref = frames_to_beat_times(r["beat_targets"][:maxT].cpu().numpy(), fps)
             ref_db = frames_to_beat_times(r["downbeat_targets"][:maxT].cpu().numpy(), fps)
